+++ __init__.py
@@ -2,6 +2,7 @@
 import io
 import os
 from pygame.locals import *
+from pygame import freetype
 import numpy as np
 
 from . import shared
@@ -24,30 +25,44 @@ from .extend import *
 #               diag = 14.3：the length of the screen diagonal (inch)
 #               angel = 1.5：visual angel of single char (degree)
 # 【返回值】pg：pygame对象，win：显示屏对象，font：字体对象
-def start(settingfile='setting.txt', fullscreen=True, normalFontSize = 20, stimFontSize = None, distance=60, diag = 23, angel = 2.5, fontColor = (255,255,255), backgroundColor = (128, 128, 128), sample_rate = 44100, bits = 16, channel=2):
+def start(settingfile='setting.txt', fullscreen=True,mouseVisible=False, normalFontSize = 20, stimFontSize = None, distance=60, diag = 23, angel = 2.5, fontColor = (255,255,255), backgroundColor = (128, 128, 128), sample_rate = 44100, bits = 16, channel=2, port='COM1'):
+    'Parameters'
     func_var = locals().copy()
-    
-    readSetting(settingfile)
-    for k in ['fullscreen','fontColor','backgroundColor','distance','diag','angel',
-              'normalFontSize','stimFontSize','sample_rate','bit','channel']:
-        if setting(k):
-            func_var[k] = eval('%s' %setting(k)[0])
+    shared.setting = readSetting(settingfile)
 
+    for k,v in shared.setting.items():
+        if k in ['fullscreen','fontColor','mouseVisible','backgroundColor','distance','diag','angel',
+              'sample_rate','bit','channel','port'] or k[-8:] == 'FontSize':
+            func_var[k] = eval('%s' %v[0])
+        else:
+            func_var[k] = eval('%s' %v)
+
+    shared.setting = func_var
+
+    'Color'
+    shared.fontColor = shared.setting['fontColor']
+    shared.backgroundColor = shared.setting['backgroundColor']
+
+    'Mouse pointer visibility'
     # Set the pointer visibility
-    shared.pg.mouse.set_visible(False)
+    shared.pg.mouse.set_visible(shared.setting['mouseVisible'])
     shared.pg.mixer.quit()
 
+    'Sound mixer'
     # Initate the mixer
-    shared.pg.mixer.init(func_var['sample_rate'], -func_var['bits'], func_var['channel'])
+    shared.pg.mixer.init(shared.setting['sample_rate'], -shared.setting['bits'], shared.setting['channel'])
     
+    'Window'
     # Initate the window
-    if func_var['fullscreen'] == True:
+    if shared.setting['fullscreen'] == True:
         shared.win = shared.pg.display.set_mode((shared.winWidth,shared.winHeight), FULLSCREEN| HWSURFACE|DOUBLEBUF)
     else:
         shared.win = shared.pg.display.set_mode((800,600), HWSURFACE | DOUBLEBUF)
         shared.winWidth = 800
         shared.winHeight = 600
+    clear() # Reset screen color
 
+    'Joystick'
     # Initate the joystick
     try:
         joystick = shared.pg.joystick.Joystick(0)
@@ -55,19 +70,25 @@ def start(settingfile='setting.txt', fullscreen=True, normalFontSize = 20, stimF
     except:
         pass
 
-    shared.fontColor = func_var['fontColor']
-    shared.backgroundColor = func_var['backgroundColor']
+    'Font'
+    # Get the font size attribute of normal text, stimulus, or others
+    for k,v in shared.setting.items():
+         if 'FontSize' == k[-8:]:
+            shared.font[k] = v
 
-    # Get the font attribute of normal text
-    shared.font['nSize'] = func_var['normalFontSize']
-    shared.font['nFont'] = shared.pg.font.Font(shared.path+"simhei.ttf", shared.font['nSize'])
+    if shared.setting['stimFontSize'] == None:
+        shared.font['stimFontSize'] =  int((shared.winWidth**2 + shared.winHeight**2) **0.5/(shared.setting['diag']*2.54 / (shared.setting['distance'] * np.tan(shared.setting['angel']/4*np.pi/180) * 2))) # pixelSize/(pixels in diagonal) = realLength/(real length in diagonal)
+    
+    for k,v in shared.font.copy().items():
+         shared.font[k[:-4]] = shared.pg.font.Font(shared.path+"simhei.ttf", shared.font[k])
 
-    # Get the font attribute of stimulus text
-    if func_var['stimFontSize'] == None:
-        shared.font['sSize'] =  int((shared.winWidth**2 + shared.winHeight**2) **0.5/(func_var['diag']*2.54 / (func_var['distance'] * np.tan(func_var['angel']/4*np.pi/180) * 2))) # pixelSize/(pixels in diagonal) = realLength/(real length in diagonal)
-    else:
-        shared.font['sSize'] =  func_var['stimFontSize']
-        
-    shared.font['sFont'] = shared.pg.font.Font(shared.path+"simhei.ttf", shared.font['sSize'])
+    shared.font['ft'] = freetype.Font(shared.path+"simhei.ttf")
 
-    clear()
+    'Port (only serial port needs port name pre-define)'
+    shared.ser.port = shared.setting['port'] # set the port
+    try:
+        shared.ser.open()
+    except:
+        print('Could not open serial port')
+
+