+++ stim/draw.py
@@ -4,93 +4,81 @@ import pygame as pg
 from expy import shared
 
 # caluate the screen position of object
-# center_x: if center_x is float, it represents the x-offset(-1~1 scale) from object center to screen center,
-#            if int, it represents the x-offset(pixel) from object center to screen upper-left.
-# center_y: similar with center_x
+# x: if x is float, it represents the x-offset(-1~1 scale) from the object benchmark to the screen center,
+#            if int, it represents the x-offset(pixel) from the object benchmark to the screen upperleft.
+# y: similar with x
 # w: if w is float, it represents the width scale on screen,
 #     if int, it represents the width in pixel.
 # h: similar with w
-def getPos(center_x=shared.winWidth//2,center_y=shared.winHeight//2,w=0,h=0):
-    if type(center_x) is float:
-        center_x, center_y = (0.5+center_x/2)*shared.winWidth, (0.5+center_y/2)*shared.winHeight
+# return position of the object's upperleft corner
+def getPos(x=shared.winWidth//2,y=shared.winHeight//2,w=0,h=0,benchmark='center'):
+    if type(x) is float:
+        x, y = (0.5+x/2)*shared.winWidth, (0.5+y/2)*shared.winHeight
     if w<1:
         w, h = w*shared.winWidth, h*shared.winHeight
-    return int(center_x-w/2),int(center_y-h/2)
+
+    if benchmark=='center':
+        return int(x-w/2),int(y-h/2)
+    elif benchmark=='upper_left':
+        return int(x),int(y)
+    elif benchmark=='upper_right':
+        return int(x-w),int(y)
+    elif benchmark=='lower_left':
+        return int(x),int(y-h)
+    elif benchmark=='lower_right':
+        return int(x-w),int(y-h)
+    elif benchmark=='upper_center':
+        return int(x-w/2),int(y)
+    elif benchmark=='left_center':
+        return int(x),int(y-h/2)
+    elif benchmark=='lower_center':
+        return int(x-w/2),int(y-h)
+    elif benchmark=='right_center':
+        return int(x-w),int(y-h/2)
+    raise ValueError('Unsupported position benchmark')
 
 '''Basic draw class'''
-# a central fixation
-# 屏幕中央显示一个红点
-def drawFix(size): 
-    x,y = getPos(w=size,h=size)
-    shared.pg.draw.circle(shared.win, (128,0,0), (x,y), size)
-    # shared.pg.draw.line(win, fixColor, (shared.winWidth / 2 - length/2, shared.winHeight / 2) , (shared.winWidth / 2 + length/2, shared.winHeight / 2), width=1)
-    # shared.pg.draw.line(win, fixColor, (shared.winWidth / 2, shared.winHeight / 2 - length/2) , (shared.winWidth / 2, shared.winHeight / 2 + length/2) , width=1)
-
-# draw word at screen center
-# 将text的内容生成并放入缓存 ，只能显示一行
-def drawWord(text, fontname='sFont', x=0.0, y=0.0):
-    target = shared.font[fontname].render(text, True, shared.fontColor)
-    x,y = getPos(x,y,w=target.get_width(), h=target.get_height())
-    shared.win.blit(target, (x, y))
-
-def drawWordWithLine(text, linePos, fontname='sFont'):
-    # shared.font[fontname].set_underline(True)
-    target = shared.font[fontname].render(text, True, shared.fontColor)
-    x,y = getPos(w=target.get_width(), h=target.get_height())
-    shared.win.blit(target, (x,y))
-
-    y = shared.winHeight / 2 + target.get_height() / 2 + 5
-    w = shared.winWidth
-    if linePos == 0:
-        linePosStart = [(w - target.get_width())/2, y]
-        linePosEnd = [(w + target.get_width())/2, y]
-    if linePos == 1:
-        linePosStart = [(w - target.get_width())/2, y]
-        linePosEnd = [w / 2, y]
-    elif linePos == 2:
-        linePosStart = [w / 2, y]
-        linePosEnd = [(w + target.get_width())/2, y]
-    shared.pg.draw.line(shared.win, shared.fontColor, linePosStart, linePosEnd, 2)
-
-# draw sentences in more than one line
+# draw text 
 # 将text的内容生成并放入缓存，text可分行显示（由‘\n’分行）
-def drawText(text, fontname='sFont', fontsize='sSize', pos='center'):
-    lines = text.split('\n')
-    lineN = len(lines)
-        
-    font = shared.font[fontname]
-    targets = [font.render(l, True, shared.fontColor) for l in lines]
-    maxLen = max([t.get_width() for t in targets])
+def drawText(text, fontname='stimFont', x=0.0, y=0.0, benchmark='center'):
+    if not '\n' in text:
+        target = shared.font[fontname].render(text, True, shared.fontColor)
+        x,y = getPos(x,y,w=target.get_width(), h=target.get_height(), benchmark=benchmark)
+        shared.win.blit(target, (x, y))
 
-    for ind,target in enumerate(targets):
-        if pos=='center':
-            x,y = getPos(w=maxLen, h=(lineN-1-ind*2)*target.get_height())
-            shared.win.blit(target, (x, y))
-        elif pos == 'LU':
-            x,y = getPos(-0.9, -0.9, 0, 0)
-            shared.win.blit(target, (x, y+ind*target.get_height()))
-        elif pos == 'RU':
-            x,y = getPos(0.9, -0.9, 0, 0)
-            shared.win.blit(target, (x-maxLen, y+ind*target.get_height()))
-        elif pos == 'LD':
-            x,y = getPos(-0.9, 0.9, 0, 0)
-            shared.win.blit(target, (x, y-(lineN-ind)*target.get_height()))
-        elif pos == 'RD':
-            x,y = getPos(0.9, 0.9, 0, 0)
-            shared.win.blit(target, (x-maxLen, y-(lineN-ind)*target.get_height()))
+    else:
+        lines = text.split('\n')
+        lineN = len(lines)
+            
+        font = shared.font[fontname]
+        targets = [font.render(l, True, shared.fontColor) for l in lines]
+        maxLen = max([t.get_width() for t in targets])
+        for ind,target in enumerate(targets):
+            y_offset = (lineN-1-ind*2)*(target.get_height()/shared.winHeight)
+            pos_x,pos_y = getPos(x,y-y_offset,w=maxLen, h=0, benchmark=benchmark)
+            shared.win.blit(target, (pos_x, pos_y))
+
+# draw text with complex format
+def drawFormattedText(text, fontname='stimFont',size=15, x=0.0, y=0.0, benchmark='center'):
+    FONT = shared.font['ft']
+    FONT.underline = True
+
+    x,y = getPos(x,y,w=size*len(text), h=size, benchmark=benchmark)
+    FONT.render_to(shared.win, (x,y), text, size=(size, size))
+
 
 # draw a square
 # 屏幕中央显示一个方块
-def drawRect(w, h, x=0.0, y=0.0, fill=True, color=(255,255,255), width=1):
-    x,y = getPos(x, y, w=w, h=h)
+def drawRect(w, h, x=0.0, y=0.0, fill=True, color=(255,255,255), width=1, benchmark='center'):
+    x,y = getPos(x, y, w=w, h=h, benchmark=benchmark)
     #colour, (x, y), size, thickness
     if fill:
         width = 0
     shared.pg.draw.rect(shared.win, color, (x, y, w, h), width)
 
 # draw circle
-def drawCircle(r, x=0.0, y=0.0, fill=True, color=(255,255,255), width=1):
-    x,y = getPos(x, y, w=0, h=0)
+def drawCircle(r, x=0.0, y=0.0, fill=True, color=(255,255,255), width=1, benchmark='center'):
+    x,y = getPos(x, y, w=0, h=0, benchmark=benchmark)
     if fill:
         width = 0
     shared.pg.draw.circle(shared.win, color, (x, y), r, width)
@@ -99,13 +87,12 @@ def drawCircle(r, x=0.0, y=0.0, fill=True, color=(255,255,255), width=1):
 def drawLine(points, color=(255,255,255), width=1):
     shared.pg.draw.lines(shared.win, color, False, points, width)
 
-
 # draw picture
-def drawPic(path, w=0, h=0, x=0.0, y=0.0):
+def drawPic(path, w=0, h=0, x=0.0, y=0.0, benchmark='center'):
     im = shared.pg.image.load(path).convert()
     if w>0 and h>0:
         im = shared.pg.transform.scale(im, (w,h))
     else:
         w,h = im.get_rect().size
-    x,y = getPos(x,y,w,h)
+    x,y = getPos(x,y,w,h, benchmark=benchmark)
     shared.win.blit(im, (x,y))