+++ test/show_picture.py
@@ -7,11 +7,17 @@ sys.path.append('../../')
 from expy import * # Import the needed functions
 start() # Initiate the experiment environment
 
+'''General usage'''
 drawPic('data/demo.jpg') # Draw a picture on the canvas center
 show(3000) # Display current canvas
+''''''
+
 
 drawPic('data/demo.jpg', w=400, h=300) # Draw a zoomed picture on the canvas center
 show(3000) # Display current canvas
 
 drawPic('data/demo.jpg', w=400, h=300, x=0.5, y=0.5) # Draw a zoomed picture on the canvas, and move it
+show(3000) # Display current canvas
+
+drawPic('data/demo.jpg', w=400, h=300, x=0.5, y=0.5, benchmark='upper_center') # Draw a zoomed picture on the canvas, and move it
 show(3000) # Display current canvas